+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Include PE SSH keys on MC machines.
 
+### Removed
+
+- Remove unused `ami_owner` variable.
+
+
 ## [3.6.0] - 2021-06-14
 
 ### Changed